[SPARK-37792][CORE] Fix the check of custom configuration in SparkShellLoggingFilter
### What changes were proposed in this pull request?

This patch fixes the way we check if the log came from root or some custom configuration in `SparkShellLoggingFilter`.

### Why are the changes needed?

Previously in log4j 1.x the way we check if the log came from root or some custom configuration in `SparkShellLoggingFilter` does not work with log4j 2 now.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Manual test

Closes #35080 from viirya/SPARK-37792.

Authored-by: Liang-Chi Hsieh <viirya@gmail.com>
Signed-off-by: Kousuke Saruta <sarutak@oss.nttdata.com>