[SPARK-37819][K8S] Add OUTLIER executor roll policy and use it by default (#1323)
### What changes were proposed in this pull request?

This PR aims to add a new executor roll policy, `OUTLIER`, which aims to detect various outliers first, and use it by default.
If there is no outlier, it will work like `TOTAL_DURATION` policy.

### Why are the changes needed?

The users can use `OUTLIER` policy to consider the outliers in terms of multiple dimensions.
In addition, this will be a better default policy.

### Does this PR introduce _any_ user-facing change?

No. This is a new feature in Apache Spark 3.3.

### How was this patch tested?

Pass the CIs with the newly added test cases.