[SPARK-37887][CORE] Fix the check of repl log level
This patch fixes the check of repl's log level. So we can correctly know if the repl class is set with log level or not.

Same as the check in `SparkShellLoggingFilter`, `getLevel` cannot be used anymore to check if the log level is set or not for a logger in log4j2.

No

Manual verified locally.

Closes #35198 from viirya/SPARK-37887.

Authored-by: Liang-Chi Hsieh <viirya@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>