[SPARK-38047][K8S] Add `OUTLIER_NO_FALLBACK` executor roll policy (#1337)
### What changes were proposed in this pull request?

This is for rdar://88338827 (Backport SPARK-38047 Add `OUTLIER_NO_FALLBACK` executor roll policy).

This PR aims to add a new executor roll policy which allows users to skip rolling in cases where there are no outlier executors.

### Why are the changes needed?

As currently implemented an executor is always rolled every `spark.kubernetes.executor.rollInterval` interval. In environments where starting of executors can introduce latencies it may be desirable for users to have the option to determine if rolling should only happen when outliers are found.

### Does this PR introduce _any_ user-facing change?

No, this is an additional option being added to a new feature in Apache Spark 3.3.

### How was this patch tested?

Pass the CIs with the newly added test cases.