[SPARK-37865][SQL] Fix union deduplication correctness bug
### What changes were proposed in this pull request?

Fixes a correctness bug in `Union` in the case that there are duplicate output columns. Previously, duplicate columns on one side of the union would result in a duplicate column being output on the other side of the union.

To do so, we go through the union’s child’s output and find the duplicates. For each duplicate set, there is a first duplicate: this one is left alone. All following duplicates are aliased and given a tag; this tag is used to remove ambiguity during resolution.

As the first duplicate is left alone, the user can still select it, avoiding a breaking change. As the later duplicates are given new expression IDs, this fixes the correctness bug.

### Why are the changes needed?

Output of union with duplicate columns in the children was incorrect

### Does this PR introduce _any_ user-facing change?

Example query:
```
SELECT a, a FROM VALUES (1, 1), (1, 2) AS t1(a, b)
UNION ALL SELECT c, d FROM VALUES (2, 2), (2, 3) AS t2(c, d)
```

Result before:
```
a | a
_ | _
1 | 1
1 | 1
2 | 2
2 | 2
```

Result after:
```
a | a
_ | _
1 | 1
1 | 1
2 | 2
2 | 3
```

### How was this patch tested?

Unit tests

Closes #35760 from karenfeng/spark-37865.

Authored-by: Karen Feng <karen.feng@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit 59ce0a706cb52a54244a747d0a070b61f5cddd1c)
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit 60e3f42210bca6d8ad4ccb268ae69bcb0dce994e)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>