rdar://92996732 Use quoted identifier for column name 'date' before pushing filter to UC (#1412)
### What changes were proposed in this pull request?

This PR aims to introduce `spark.sql.hive.metastorePartitionQuoting` configuration to quote 
a column when its name is `date` which is reserved keyword in UC.

### Why are the changes needed?

This is specific to Aurora project which is migrating ancient HMS to UC.

### Does this PR introduce _any_ user-facing change?

No. This is disabled by default.

### How was this patch tested?

The test case verified that new configuration doesn't break the other columns.

The new configuration should be tested against UC and Aurora applications on their schema having `date` column.