[SPARK-39134][SS] Add custom metric of skipped null values for stream join operator (#1414)
### What changes were proposed in this pull request?

This proposes to add a custom metric of skipped null values for stream-stream join operator.

### Why are the changes needed?

When `STATE_STORE_SKIP_NULLS_FOR_STREAM_STREAM_JOINS` is enabled, Spark will skip null values in hash-based stream-stream joins. But currently we cannot track if this is happened and how many it is. For our customers, they have an interest to know the number of values skipped for better tracking purpose.

### Does this PR introduce _any_ user-facing change?

Yes. One custom metric is added to stream-stream join if the config is enabled. As the config is disabled by default, it is 'no' for users by default.

### How was this patch tested?

Updated current test.

Closes #36491 from viirya/add_skipped_null_count_metric.

Authored-by: Liang-Chi Hsieh <viirya@gmail.com>
Signed-off-by: Liang-Chi Hsieh <viirya@gmail.com>

Co-authored-by: Liang-Chi Hsieh <viirya@gmail.com>