rdar://93691429 Check Merge Alignment Referenced with correct Case Sensitivity (#1428)
### What changes were proposed in this pull request?
Previously alignment was checked by comparing the exact attribute references between
Spark and the underlying table, which failed with case insensitive SQL configurations. To
fix this we use the configuration's resolver to compare references.

### Why are the changes needed?
This was breaking some migrations from Spark 3.1 where the alignment check was not present. A query which attempted
to do a MergeInto with column names which matched in a case-insensitive way would fail to trigger our plan rewrite
rules leading to an opaque MERGE INTO is temporarily not supported exception.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
This patch is a cherry pick of the code fixed and released in Apache Iceberg. https://github.com/apache/iceberg/pull/4848 

The test for this specific case is in the Iceberg codebase and we will back-port this to ADT following the merge of the fix in Spark.