[SPARK-39340][SQL][3.2] DS v2 agg pushdown should allow dots in the name of top-level columns
### What changes were proposed in this pull request?

In the first version of DS v2 aggregate pushdown, we don't want to support nested fields and we picked `PushableColumnWithoutNestedColumn` to match top-level columns. However, this introduced an unexpected limitation: the column name cannot contain dots. This makes sense for DS v2 filter pushdown to keep backward compatibility, but DS v2 agg pushdown is a completely new feature and we don't need to worry about backward compatibility.

This PR removes this limitation. Note that in Spark 3.3, DS v2 agg pushdown supports nested fields and this limitation is gone.

### Why are the changes needed?

Unexpected limitation is a mistake

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

new tests

Closes #36727 from cloud-fan/agg.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>