rdar://100081966: Add Rio pipeline to run iceberg unit tests (#1573)
We recently hit by an issue due to a Hive upgrade doesn’t work with Iceberg. As Apple Spark is heavily used with Iceberg in the production, any change at Spark has a risk to affect Iceberg function. But we don’t run any tests against Iceberg at the moment.

To prevent similar issue on Iceberg side, it would be nice if we can run Iceberg unit tests in Apple Spark Rio pipeline.