[SPARK-42188][BUILD][3.2] Force SBT protobuf version to match Maven
Update `SparkBuild.scala` to force SBT use of `protobuf-java` to match the Maven version.  The Maven dependencyManagement section forces `protobuf-java` to use `2.5.0`, but SBT is using `3.14.0`.

Define `protoVersion` in `SparkBuild.scala` and use it in `DependencyOverrides` to force the SBT version of `protobuf-java` to match the setting defined in the Maven top-level `pom.xml`.  Add comments to both `pom.xml` and `SparkBuild.scala` to ensure that the values are kept in sync.

No

Before the update, SBT reported using `3.14.0`:
```
% build/sbt dependencyTree | grep proto | sed 's/^.*-com/com/' | sort | uniq -c
   8 com.google.protobuf:protobuf-java:2.5.0 (evicted by: 3.14.0)
  70 com.google.protobuf:protobuf-java:3.14.0
```

After the patch is applied, SBT reports using `2.5.0`:
```
% build/sbt dependencyTree | grep proto | sed 's/^.*-com/com/' | sort | uniq -c
  70 com.google.protobuf:protobuf-java:2.5.0
```

Closes #39745 from snmvaughan/feature/SPARK-42188.

Authored-by: Steve Vaughan Jr <s_vaughan@apple.com>
Signed-off-by: huaxingao <huaxin_gao@apple.com>