rdar://106603368: [Boson] build: replace Boson dep version in Iceberg test in Spark rio (#1707)
We run Iceberg unit tests in Spark rio. During it, it replaces Iceberg’s Hive and Spark version with latest versions from checkout Spark repo, to ensure latest Spark/Hive work with Iceberg.

As Boson is included in both Iceberg and Spark like Hive, we need to do the same for Boson dependency.