[SPARK-42799][BUILD] Update SBT build `xercesImpl` version to match with `pom.xml`
This PR aims to update `XercesImpl` version to `2.12.2` from `2.12.0` in order to match with the version of `pom.xml`.

https://github.com/apache/spark/blob/149e020a5ca88b2db9c56a9d48e0c1c896b57069/pom.xml#L1429-L1433

When we updated this version via SPARK-39183, we missed to update `SparkBuild.scala`.

- https://github.com/apache/spark/pull/36544

No, this is a dev-only change because the release artifact' dependency is managed by Maven.

Pass the CIs.

Closes #40431 from dongjoon-hyun/SPARK-42799.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 049aa380b8b1361c2898bc499e64613d329c6f72)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>